requirements/prod.txt
django
djangorestframework
djangorestframework-simplejwt
markdown
gunicorn
whitenoise
